+++ cartoframes/vis/style.py
@@ -1,6 +1,7 @@
 from __future__ import absolute_import
 
-from . import defaults
+from .defaults import STYLE_DEFAULTS, STYLE_PROPERTIES
+from ..utils import merge_dicts
 
 
 class Style(object):
@@ -51,43 +52,48 @@ class Style(object):
 
     def _init_style(self, data):
         if data is None:
-            return defaults.STYLE_DEFAULTS
+            return STYLE_DEFAULTS
         elif isinstance(data, (str, dict)):
             return data
         else:
             raise ValueError('`style` must be a string or a dictionary')
 
-    def compute_viz(self, geom_type=None, variables={}):
+    def compute_viz(self, geom_type, variables={}):
         style = self._style
+        defaults = STYLE_DEFAULTS[geom_type]
         if isinstance(style, dict):
-            if geom_type and geom_type in style:
+            if geom_type in style:
                 style = style.get(geom_type)
-            return self._parse_style_dict(style, variables)
+            return self._parse_style_dict(style, defaults, variables)
         elif isinstance(style, str):
-            return self._parse_style_str(style, variables)
+            return self._parse_style_str(style, defaults, variables)
         else:
             raise ValueError('`style` must be a string or a dictionary')
 
-    def _parse_style_dict(self, style, ext_vars):
-        style_vars = style.get('vars', {})
-        variables = _merge_dicts(style_vars, ext_vars)
+    def _parse_style_dict(self, style, defaults, ext_vars):
+        variables = merge_dicts(style.get('vars', {}), ext_vars)
+        properties = merge_dicts(defaults, style)
 
         serialized_variables = self._serialize_variables(variables)
-        serialized_properties = self._serialize_properties(style)
+        serialized_properties = self._serialize_properties(properties)
 
         return serialized_variables + serialized_properties
 
-    def _parse_style_str(self, style, ext_vars):
-        serialized_variables = self._serialize_variables(ext_vars)
+    def _parse_style_str(self, style, defaults, ext_vars):
+        variables = ext_vars
+        default_properties = self._prune_defaults(defaults, style)
 
-        return serialized_variables + style
+        serialized_variables = self._serialize_variables(variables)
+        serialized_default_properties = self._serialize_properties(default_properties)
+
+        return serialized_variables + serialized_default_properties + style
 
     def _serialize_variables(self, variables={}):
         output = ''
         for var in variables:
             output += '@{name}: {value}\n'.format(
                 name=var,
-                value=_convstr(variables.get(var))
+                value=variables.get(var)
             )
         return output
 
@@ -96,25 +102,26 @@ class Style(object):
         for prop in properties:
             if prop == 'vars':
                 continue
-            if prop not in defaults.STYLE_PROPERTIES:
+            if prop not in STYLE_PROPERTIES:
                 raise ValueError(
                     'Style property "{0}" is not valid. Valid style properties are: {1}'.format(
                         prop,
-                        ', '.join(defaults.STYLE_PROPERTIES)
+                        ', '.join(STYLE_PROPERTIES)
                     ))
             output += '{name}: {value}\n'.format(
                 name=prop,
-                value=_convstr(properties.get(prop))
+                value=properties.get(prop)
             )
         return output
 
-
-def _convstr(obj):
-    """Converts all types to strings or None"""
-    return str(obj) if obj is not None else None
-
-
-def _merge_dicts(dict1, dict2):
-    d = dict1.copy()
-    d.update(dict2)
-    return d
+    def _prune_defaults(self, defaults, style):
+        output = defaults.copy()
+        if 'color:' in style:
+            del output['color']
+        if 'width:' in style:
+            del output['width']
+        if 'strokeWidth:' in style:
+            del output['strokeWidth']
+        if 'strokeColor:' in style:
+            del output['strokeColor']
+        return output