+++ test/test_utils.py
@@ -11,39 +11,40 @@ from cartoframes.utils import (dict_items, cssify, importify_params)
 
 class TestUtils(unittest.TestCase):
     """Tests for functions in utils module"""
+
     def setUp(self):
         self.point_style = {
             "#layer['mapnik::geometry_type'=1]": OrderedDict([
-                                              ('marker-width', "6"),
-                                              ('marker-fill', "yellow"),
-                                              ('marker-fill-opacity', "1"),
-                                              ('marker-allow-overlap', "true"),
-                                              ('marker-line-width', "0.5"),
-                                              ('marker-line-color', "black"),
-                                              ("marker-line-opacity", "1")])
-            }
+                ('marker-width', "6"),
+                ('marker-fill', "yellow"),
+                ('marker-fill-opacity', "1"),
+                ('marker-allow-overlap', "true"),
+                ('marker-line-width', "0.5"),
+                ('marker-line-color', "black"),
+                ("marker-line-opacity", "1")])
+        }
 
         self.polygon_style = {
-                "#layer['mapnik::geometry_type'=3]": OrderedDict([
-                    ('polygon-fill', ('ramp([column], (#ffc6c4, #ee919b, '
-                                      '#cc607d, #9e3963, #672044), '
-                                      'quantiles)')),
-                    ('polygon-opacity', '0.9'),
-                    ('polygon-gamma', '0.5'),
-                    ('line-color', '#FFF'),
-                    ('line-width', '0.5'),
-                    ('line-opacity', '0.25'),
-                    ('line-comp-op', 'hard-light')])}
+            "#layer['mapnik::geometry_type'=3]": OrderedDict([
+                ('polygon-fill', ('ramp([column], (#ffc6c4, #ee919b, '
+                                  '#cc607d, #9e3963, #672044), '
+                                  'quantiles)')),
+                ('polygon-opacity', '0.9'),
+                ('polygon-gamma', '0.5'),
+                ('line-color', '#FFF'),
+                ('line-width', '0.5'),
+                ('line-opacity', '0.25'),
+                ('line-comp-op', 'hard-light')])}
 
         self.complex_style = OrderedDict([
             ("#layer['mapnik::geometry_type'=1]", OrderedDict([
-                                              ('marker-width', "5"),
-                                              ('marker-fill', "yellow"),
-                                              ('marker-fill-opacity', '1'),
-                                              ('marker-allow-overlap', 'true'),
-                                              ('marker-line-width', '0.5'),
-                                              ('marker-line-color', "black"),
-                                              ('marker-line-opacity', '1')])),
+                ('marker-width', "5"),
+                ('marker-fill', "yellow"),
+                ('marker-fill-opacity', '1'),
+                ('marker-allow-overlap', 'true'),
+                ('marker-line-width', '0.5'),
+                ('marker-line-color', "black"),
+                ('marker-line-opacity', '1')])),
             ("#layer['mapnik::geometry_type'=2]", OrderedDict([
                 ('line-width', '1.5'),
                 ('line-color', "black")])),