+++ test/utils.py
@@ -3,6 +3,7 @@ import os
 import json
 import warnings
 
+
 class _UserUrlLoader:
     def user_url(self):
         user_url = None